+++ src/components/dac_dataset_table/DACDatasetTableCellData.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import style from 'src/pages/DACDatasets.module.css';
 import {styles} from './DACDatasetConstants';
 import DACDatasetApprovalStatus from './DACDatasetApprovalStatus';
-import ReactTooltip from 'react-tooltip';
+import {createDataUseDisplay, processDataUseCodes} from 'src/utils/DataUseUtils.js';
 
 export const consoleTypes = { CHAIR: 'chair' };
 
@@ -69,71 +69,6 @@ export function dataCustodianCellData({dataset, label = 'dataCustodianCellData'}
     cellStyle: {width: styles.cellWidths.dataCustodian},
     label
   };
-}
-
-/**
- * Process data use codes from a dataset
- * @param {Object} dataset - DatasetTerm object containing data use information
- * @returns {Object} - Object with processed data use information
- */
-export function processDataUseCodes(dataset) {
-  const codesAndDescriptions = dataset.dataUse?.primary ? dataset.dataUse.primary.map((dataUse) => {
-    if (dataUse.code === 'OTHER') {
-      return {'code': `OTH1`, 'description': dataUse.description};
-    } else if (dataUse.code === 'DS') {
-      const disease = dataUse.description.substring(dataUse.description.indexOf(':') + 2);
-      return {'code': `${dataUse.code} (${disease})`, 'description': dataUse.description};
-    } else {
-      return {'code': dataUse.code, 'description': dataUse.description};
-    }
-  }) : [];
-
-  if (dataset.dataUse?.secondary) {
-    dataset.dataUse.secondary.forEach((dataUse) => {
-      if (dataUse.code === 'OTHER') {
-        codesAndDescriptions.push({'code': `OTH2`, 'description': dataUse.description});
-      } else {
-        codesAndDescriptions.push({'code': dataUse.code, 'description': dataUse.description});
-      }
-    });
-  }
-
-  const codeList = codesAndDescriptions.map(du => du.code);
-
-  return { codesAndDescriptions, codeList };
-}
-
-/**
- * Creates a data use display component with tooltips
- * @param {Object} dataset - DatasetTerm object
- * @param {string} divClass - CSS class for the div
- * @param {string} spanClass - CSS class for the span
- * @param {string} tooltipPlace - Placement direction for tooltip
- * @returns {JSX.Element} - Element displaying data use codes with tooltips
- */
-export function createDataUseDisplay({
-                                       dataset,
-                                       divClass,
-                                       spanClass,
-                                       tooltipPlace = 'right'
-                                     }) {
-  const { codesAndDescriptions, codeList } = processDataUseCodes(dataset);
-
-  return (
-      <div style={{display: 'inline-block'}} className={divClass}>
-      <span className={spanClass} data-tip={true} data-for={`dataset-data-use-${dataset.datasetId}`}>
-        {codeList.join(', ')}
-      </span>
-        <ReactTooltip
-            place={tooltipPlace}
-            effect={'solid'}
-            id={`dataset-data-use-${dataset.datasetId}`}>
-          <ul>{codesAndDescriptions.map((translation, index) => {
-            return <li key={`${translation.code}_s_${index}`}>{translation.code}: {translation.description}</li>;
-          })}</ul>
-        </ReactTooltip>
-      </div>
-  );
 }
 
 export function dataUseCellData({dataset, label = 'dataUseCellData', divClass = style['cell-data'], spanClass = style['data-use'], cellWidth = styles.cellWidths.dataUse, tooltipPlace = 'right'}) {
